chore: remove DocumentContextProvider
