+++ packages/core/content-manager/admin/src/pages/EditView/EditViewPage.tsx
@@ -15,7 +15,6 @@ import { styled } from 'styled-components';
 
 import { SINGLE_TYPES } from '../../constants/collections';
 import { PERMISSIONS } from '../../constants/plugin';
-import { DocumentContextProvider } from '../../features/DocumentContext';
 import { DocumentRBAC, useDocumentRBAC } from '../../features/DocumentRBAC';
 import { useDoc, type UseDocument } from '../../hooks/useDocument';
 import { useDocumentLayout } from '../../hooks/useDocumentLayout';
@@ -128,83 +127,75 @@ const EditViewPage = () => {
   return (
     <Main paddingLeft={10} paddingRight={10}>
       <Page.Title>{getTitle(mainField)}</Page.Title>
-      <DocumentContextProvider
-        initialDocument={{
-          documentId: id!,
-          model,
-          collectionType,
-        }}
-      >
-        <Form
-          disabled={hasDraftAndPublished && status === 'published'}
-          initialValues={initialValues}
-          method={isCreatingDocument ? 'POST' : 'PUT'}
-          validate={(values: Record<string, unknown>, options: Record<string, string>) => {
-            const yupSchema = createYupSchema(schema?.attributes, components, {
-              status,
-              ...options,
-            });
+      <Form
+        disabled={hasDraftAndPublished && status === 'published'}
+        initialValues={initialValues}
+        method={isCreatingDocument ? 'POST' : 'PUT'}
+        validate={(values: Record<string, unknown>, options: Record<string, string>) => {
+          const yupSchema = createYupSchema(schema?.attributes, components, {
+            status,
+            ...options,
+          });
 
-            return yupSchema.validate(values, { abortEarly: false });
-          }}
-          initialErrors={location?.state?.forceValidation ? validateSync(initialValues, {}) : {}}
-        >
-          {({ resetForm }) => (
-            <>
-              <Header
-                isCreating={isCreatingDocument}
-                status={hasDraftAndPublished ? getDocumentStatus(document, meta) : undefined}
-                title={getTitle(mainField)}
-              />
-              <Tabs.Root variant="simple" value={status} onValueChange={handleTabChange}>
-                <Tabs.List
-                  aria-label={formatMessage({
-                    id: getTranslation('containers.edit.tabs.label'),
-                    defaultMessage: 'Document status',
-                  })}
-                >
-                  {hasDraftAndPublished ? (
-                    <>
-                      <StatusTab value="draft">
-                        {formatMessage({
-                          id: getTranslation('containers.edit.tabs.draft'),
-                          defaultMessage: 'draft',
-                        })}
-                      </StatusTab>
-                      <StatusTab
-                        disabled={!meta || meta.availableStatus.length === 0}
-                        value="published"
-                      >
-                        {formatMessage({
-                          id: getTranslation('containers.edit.tabs.published'),
-                          defaultMessage: 'published',
-                        })}
-                      </StatusTab>
-                    </>
-                  ) : null}
-                </Tabs.List>
-                <Grid.Root paddingTop={8} gap={4}>
-                  <Grid.Item col={9} s={12} direction="column" alignItems="stretch">
-                    <Tabs.Content value="draft">
-                      <FormLayout layout={layout} document={doc} />
-                    </Tabs.Content>
-                    <Tabs.Content value="published">
-                      <FormLayout layout={layout} document={doc} />
-                    </Tabs.Content>
-                  </Grid.Item>
-                  <Grid.Item col={3} s={12} direction="column" alignItems="stretch">
-                    <Panels />
-                  </Grid.Item>
-                </Grid.Root>
-              </Tabs.Root>
-              <Blocker
-                // We reset the form to the published version to avoid errors like – https://strapi-inc.atlassian.net/browse/CONTENT-2284
-                onProceed={resetForm}
-              />
-            </>
-          )}
-        </Form>
-      </DocumentContextProvider>
+          return yupSchema.validate(values, { abortEarly: false });
+        }}
+        initialErrors={location?.state?.forceValidation ? validateSync(initialValues, {}) : {}}
+      >
+        {({ resetForm }) => (
+          <>
+            <Header
+              isCreating={isCreatingDocument}
+              status={hasDraftAndPublished ? getDocumentStatus(document, meta) : undefined}
+              title={getTitle(mainField)}
+            />
+            <Tabs.Root variant="simple" value={status} onValueChange={handleTabChange}>
+              <Tabs.List
+                aria-label={formatMessage({
+                  id: getTranslation('containers.edit.tabs.label'),
+                  defaultMessage: 'Document status',
+                })}
+              >
+                {hasDraftAndPublished ? (
+                  <>
+                    <StatusTab value="draft">
+                      {formatMessage({
+                        id: getTranslation('containers.edit.tabs.draft'),
+                        defaultMessage: 'draft',
+                      })}
+                    </StatusTab>
+                    <StatusTab
+                      disabled={!meta || meta.availableStatus.length === 0}
+                      value="published"
+                    >
+                      {formatMessage({
+                        id: getTranslation('containers.edit.tabs.published'),
+                        defaultMessage: 'published',
+                      })}
+                    </StatusTab>
+                  </>
+                ) : null}
+              </Tabs.List>
+              <Grid.Root paddingTop={8} gap={4}>
+                <Grid.Item col={9} s={12} direction="column" alignItems="stretch">
+                  <Tabs.Content value="draft">
+                    <FormLayout layout={layout} document={doc} />
+                  </Tabs.Content>
+                  <Tabs.Content value="published">
+                    <FormLayout layout={layout} document={doc} />
+                  </Tabs.Content>
+                </Grid.Item>
+                <Grid.Item col={3} s={12} direction="column" alignItems="stretch">
+                  <Panels />
+                </Grid.Item>
+              </Grid.Root>
+            </Tabs.Root>
+            <Blocker
+              // We reset the form to the published version to avoid errors like – https://strapi-inc.atlassian.net/browse/CONTENT-2284
+              onProceed={resetForm}
+            />
+          </>
+        )}
+      </Form>
     </Main>
   );
 };
