+++ packages/core/content-manager/admin/src/pages/EditView/components/FormInputs/Component/Input.tsx
@@ -5,7 +5,7 @@ import { Field, Flex, IconButton } from '@strapi/design-system';
 import { Trash } from '@strapi/icons';
 import { useIntl } from 'react-intl';
 
-import { useNewDocumentContext } from '../../../../../features/DocumentContext';
+import { useDocumentContext } from '../../../../../hooks/useDocumentContext';
 import { EditFieldLayout } from '../../../../../hooks/useDocumentLayout';
 import { getTranslation } from '../../../../../utils/translations';
 import { transformDocument } from '../../../utils/data';
@@ -44,7 +44,7 @@ const ComponentInput = ({
 
   const {
     currentDocument: { components },
-  } = useNewDocumentContext('ComponentInput');
+  } = useDocumentContext('ComponentInput');
 
   const handleInitialisationClick = () => {
     const schema = components[attribute.component];
