+++ packages/core/content-manager/admin/src/pages/EditView/components/FormInputs/Component/NonRepeatable.tsx
@@ -2,7 +2,7 @@ import { useField } from '@strapi/admin/strapi-admin';
 import { Box, Flex } from '@strapi/design-system';
 import { useIntl } from 'react-intl';
 
-import { useNewDocumentContext } from '../../../../../features/DocumentContext';
+import { useDocumentContext } from '../../../../../hooks/useDocumentContext';
 import { ResponsiveGridItem, ResponsiveGridRoot } from '../../FormLayout';
 import { ComponentProvider, useComponent } from '../ComponentContext';
 
@@ -20,7 +20,7 @@ const NonRepeatableComponent = ({
   const { value } = useField(name);
   const level = useComponent('NonRepeatableComponent', (state) => state.level);
   const isNested = level > 0;
-  const { currentDocument } = useNewDocumentContext('NonRepeatableComponent');
+  const { currentDocument } = useDocumentContext('NonRepeatableComponent');
 
   return (
     <ComponentProvider id={value?.id} uid={attribute.component} level={level + 1} type="component">
