+++ packages/core/content-manager/admin/src/pages/EditView/components/FormInputs/DynamicZone/tests/Field.test.tsx
@@ -2,19 +2,9 @@ import { Form } from '@strapi/admin/strapi-admin';
 import { act, render as renderRTL, screen, waitFor } from '@tests/utils';
 import { Route, Routes } from 'react-router-dom';
 
-import { DocumentContextProvider } from '../../../../../../features/DocumentContext';
 import { DynamicZone, DynamicZoneProps } from '../Field';
 
 const TEST_NAME = 'DynamicZoneComponent';
-
-const relationContext = {
-  initialDocument: {
-    documentId: 'abcdefg',
-    model: 'api::address.address',
-    collectionType: 'collection-types',
-  },
-  setCurrentDocument: jest.fn(),
-};
 
 /**
  * We _could_ unmock this and use it, but it requires more
@@ -54,7 +44,7 @@ describe('DynamicZone', () => {
               path="/content-manager/:collectionType/:slug/:id"
               element={
                 <Form initialValues={initialFormValues} method="POST" onSubmit={jest.fn()}>
-                  <DocumentContextProvider {...relationContext}>{children}</DocumentContextProvider>
+                  {children}
                 </Form>
               }
             />
