+++ packages/core/content-manager/admin/src/pages/EditView/components/FormInputs/Relations/tests/RelationModal.test.tsx
@@ -1,16 +1,7 @@
 import { render, screen, fireEvent } from '@tests/utils';
 import { useNavigate } from 'react-router-dom';
 
-import { DocumentContextProvider } from '../../../../../../features/DocumentContext';
 import { RelationCard } from '../RelationModal';
-
-const relationContext = {
-  initialDocument: {
-    documentId: 'abcdefg',
-    model: 'api::test.test',
-    collectionType: 'collection-types',
-  },
-};
 
 jest.mock('react-router-dom', () => ({
   ...jest.requireActual('react-router-dom'),
@@ -215,31 +206,19 @@ describe('<RelationModal />', () => {
   });
 
   it('renders the trigger button correctly', () => {
-    render(
-      <DocumentContextProvider {...relationContext}>
-        <RelationCard triggerButtonLabel="Open Modal" relation={relation} />
-      </DocumentContextProvider>
-    );
+    render(<RelationCard triggerButtonLabel="Open Modal" relation={relation} />);
 
     expect(screen.getByText('Open Modal')).toBeInTheDocument();
   });
 
   it('does not render the modal by default', () => {
-    render(
-      <DocumentContextProvider {...relationContext}>
-        <RelationCard triggerButtonLabel="Open Modal" relation={relation} />
-      </DocumentContextProvider>
-    );
+    render(<RelationCard triggerButtonLabel="Open Modal" relation={relation} />);
 
     expect(screen.queryByRole('dialog')).not.toBeInTheDocument();
   });
 
   it('opens the modal when clicking the trigger button', () => {
-    render(
-      <DocumentContextProvider {...relationContext}>
-        <RelationCard triggerButtonLabel="Open Modal" relation={relation} />
-      </DocumentContextProvider>
-    );
+    render(<RelationCard triggerButtonLabel="Open Modal" relation={relation} />);
 
     const button = screen.getByText('Open Modal');
     fireEvent.click(button);
@@ -257,11 +236,7 @@ describe('<RelationModal />', () => {
   });
 
   it('closes the modal when clicking the cancel button', () => {
-    render(
-      <DocumentContextProvider {...relationContext}>
-        <RelationCard triggerButtonLabel="Open Modal" relation={relation} />
-      </DocumentContextProvider>
-    );
+    render(<RelationCard triggerButtonLabel="Open Modal" relation={relation} />);
 
     const button = screen.getByText('Open Modal');
     fireEvent.click(button);
@@ -275,11 +250,7 @@ describe('<RelationModal />', () => {
   });
 
   it('navigates to full page when "Go to entry" is clicked', () => {
-    render(
-      <DocumentContextProvider {...relationContext}>
-        <RelationCard triggerButtonLabel="Open Modal" relation={relation} />
-      </DocumentContextProvider>
-    );
+    render(<RelationCard triggerButtonLabel="Open Modal" relation={relation} />);
 
     fireEvent.click(screen.getByText('Open Modal'));
