+++ packages/core/content-manager/admin/src/preview/pages/Preview.tsx
@@ -16,7 +16,6 @@ import { styled } from 'styled-components';
 
 import { GetPreviewUrl } from '../../../../shared/contracts/preview';
 import { COLLECTION_TYPES } from '../../constants/collections';
-import { DocumentContextProvider } from '../../features/DocumentContext';
 import { DocumentRBAC } from '../../features/DocumentRBAC';
 import { type UseDocument, useDocument } from '../../hooks/useDocument';
 import { type EditLayout, useDocumentLayout } from '../../hooks/useDocumentLayout';
@@ -170,119 +169,110 @@ const PreviewPage = () => {
           }
         )}
       </Page.Title>
-      <DocumentContextProvider
-        initialDocument={{
-          documentId: documentId || '',
-          model,
-          collectionType,
-        }}
+      <PreviewProvider
+        url={previewUrl}
+        document={documentResponse.document}
+        title={documentTitle}
+        meta={documentResponse.meta}
+        schema={documentResponse.schema}
+        layout={documentLayoutResponse.edit}
         onPreview={onPreview}
       >
-        <PreviewProvider
-          url={previewUrl}
-          document={documentResponse.document}
-          title={documentTitle}
-          meta={documentResponse.meta}
-          schema={documentResponse.schema}
-          layout={documentLayoutResponse.edit}
-          onPreview={onPreview}
-        >
-          <FormContext
-            method="PUT"
-            disabled={
-              query.status === 'published' &&
-              documentResponse &&
-              documentResponse.document.status === 'published'
-            }
-            initialValues={documentResponse.getInitialFormValues()}
-            initialErrors={location?.state?.forceValidation ? validateSync(initialValues, {}) : {}}
-            height="100%"
-            validate={(values: Record<string, unknown>, options: Record<string, string>) => {
-              const yupSchema = createYupSchema(
-                documentResponse.schema?.attributes,
-                documentResponse.components,
-                {
-                  status: documentResponse.document?.status,
-                  ...options,
-                }
-              );
+        <FormContext
+          method="PUT"
+          disabled={
+            query.status === 'published' &&
+            documentResponse &&
+            documentResponse.document.status === 'published'
+          }
+          initialValues={documentResponse.getInitialFormValues()}
+          initialErrors={location?.state?.forceValidation ? validateSync(initialValues, {}) : {}}
+          height="100%"
+          validate={(values: Record<string, unknown>, options: Record<string, string>) => {
+            const yupSchema = createYupSchema(
+              documentResponse.schema?.attributes,
+              documentResponse.components,
+              {
+                status: documentResponse.document?.status,
+                ...options,
+              }
+            );
 
-              return yupSchema.validate(values, { abortEarly: false });
-            }}
-          >
-            {({ resetForm }) => (
-              <Flex direction="column" height="100%" alignItems="stretch">
-                <Blocker onProceed={resetForm} />
-                <PreviewHeader />
-                <Flex flex={1} overflow="auto" alignItems="stretch">
+            return yupSchema.validate(values, { abortEarly: false });
+          }}
+        >
+          {({ resetForm }) => (
+            <Flex direction="column" height="100%" alignItems="stretch">
+              <Blocker onProceed={resetForm} />
+              <PreviewHeader />
+              <Flex flex={1} overflow="auto" alignItems="stretch">
+                <Box
+                  overflow="auto"
+                  width={isSideEditorOpen ? '50%' : 0}
+                  borderWidth="0 1px 0 0"
+                  borderColor="neutral150"
+                  paddingTop={6}
+                  paddingBottom={6}
+                  // Remove horizontal padding when the editor is closed or it won't fully disappear
+                  paddingLeft={isSideEditorOpen ? 6 : 0}
+                  paddingRight={isSideEditorOpen ? 6 : 0}
+                  transition="all 0.2s ease-in-out"
+                >
+                  {/* TODO: add license check */}
+                  <FormLayout
+                    layout={documentLayoutResponse.edit.layout}
+                    document={documentResponse}
+                    hasBackground={false}
+                  />
+                </Box>
+                <Box position="relative" flex={1} height="100%" overflow="hidden">
                   <Box
-                    overflow="auto"
-                    width={isSideEditorOpen ? '50%' : 0}
-                    borderWidth="0 1px 0 0"
-                    borderColor="neutral150"
-                    paddingTop={6}
-                    paddingBottom={6}
-                    // Remove horizontal padding when the editor is closed or it won't fully disappear
-                    paddingLeft={isSideEditorOpen ? 6 : 0}
-                    paddingRight={isSideEditorOpen ? 6 : 0}
-                    transition="all 0.2s ease-in-out"
+                    data-testid="preview-iframe"
+                    ref={iframeRef}
+                    src={previewUrl}
+                    /**
+                     * For some reason, changing an iframe's src tag causes the browser to add a new item in the
+                     * history stack. This is an issue for us as it means clicking the back button will not let us
+                     * go back to the edit view. To fix it, we need to trick the browser into thinking this is a
+                     * different iframe when the preview URL changes. So we set a key prop to force React
+                     * to mount a different node when the src changes.
+                     */
+                    key={previewUrl}
+                    title={formatMessage({
+                      id: 'content-manager.preview.panel.title',
+                      defaultMessage: 'Preview',
+                    })}
+                    width="100%"
+                    height="100%"
+                    borderWidth={0}
+                    tag="iframe"
+                  />
+                  <IconButton
+                    variant="tertiary"
+                    label={formatMessage(
+                      isSideEditorOpen
+                        ? {
+                            id: 'content-manager.preview.content.close-editor',
+                            defaultMessage: 'Close editor',
+                          }
+                        : {
+                            id: 'content-manager.preview.content.open-editor',
+                            defaultMessage: 'Open editor',
+                          }
+                    )}
+                    onClick={() => setIsSideEditorOpen((prev) => !prev)}
+                    position="absolute"
+                    top={2}
+                    left={2}
                   >
-                    {/* TODO: add license check */}
-                    <FormLayout
-                      layout={documentLayoutResponse.edit.layout}
-                      document={documentResponse}
-                      hasBackground={false}
-                    />
-                  </Box>
-                  <Box position="relative" flex={1} height="100%" overflow="hidden">
-                    <Box
-                      data-testid="preview-iframe"
-                      ref={iframeRef}
-                      src={previewUrl}
-                      /**
-                       * For some reason, changing an iframe's src tag causes the browser to add a new item in the
-                       * history stack. This is an issue for us as it means clicking the back button will not let us
-                       * go back to the edit view. To fix it, we need to trick the browser into thinking this is a
-                       * different iframe when the preview URL changes. So we set a key prop to force React
-                       * to mount a different node when the src changes.
-                       */
-                      key={previewUrl}
-                      title={formatMessage({
-                        id: 'content-manager.preview.panel.title',
-                        defaultMessage: 'Preview',
-                      })}
-                      width="100%"
-                      height="100%"
-                      borderWidth={0}
-                      tag="iframe"
-                    />
-                    <IconButton
-                      variant="tertiary"
-                      label={formatMessage(
-                        isSideEditorOpen
-                          ? {
-                              id: 'content-manager.preview.content.close-editor',
-                              defaultMessage: 'Close editor',
-                            }
-                          : {
-                              id: 'content-manager.preview.content.open-editor',
-                              defaultMessage: 'Open editor',
-                            }
-                      )}
-                      onClick={() => setIsSideEditorOpen((prev) => !prev)}
-                      position="absolute"
-                      top={2}
-                      left={2}
-                    >
-                      <AnimatedArrow isSideEditorOpen={isSideEditorOpen} />
-                    </IconButton>
-                  </Box>
-                </Flex>
+                    <AnimatedArrow isSideEditorOpen={isSideEditorOpen} />
+                  </IconButton>
+                </Box>
               </Flex>
-            )}
-          </FormContext>
-        </PreviewProvider>
-      </DocumentContextProvider>
+            </Flex>
+          )}
+        </FormContext>
+      </PreviewProvider>
     </>
   );
 };
